# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PaymentSubmissionStatus = typeof PaymentSubmissionStatus[keyof typeof PaymentSubmissionStatus];


export const PaymentSubmissionStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface PaymentSubmission {
  id: number;
  orderId: string;
  paymentMethod: string;
  total: number;
  transactionLast4: string;
  status: PaymentSubmissionStatus;
  createdAt: string;
  updatedAt: string;
}

export interface PaymentSubmissionInput {
  /** @minLength 1 */
  orderId: string;
  /** @minLength 1 */
  paymentMethod: string;
  /** @minimum 0 */
  total: number;
  /** @pattern ^\d{4}$ */
  transactionLast4: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListAdminPaymentSubmissionsParams = {
status?: ListAdminPaymentSubmissionsStatus;
};

export type ListAdminPaymentSubmissionsStatus = typeof ListAdminPaymentSubmissionsStatus[keyof typeof ListAdminPaymentSubmissionsStatus];


export const ListAdminPaymentSubmissionsStatus = {
  pending: 'pending',
  paid: 'paid',
  all: 'all',
} as const;