# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: payments
    description: Customer payment confirmation operations
  - name: admin
    description: Admin payment review operations
  - name: drivers
    description: Admin driver account operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /payment-submissions:
    post:
      operationId: submitPayment
      tags: [payments]
      summary: Submit payment transaction suffix
      description: Creates a pending payment submission using the final four digits of the transaction number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentSubmissionInput"
      responses:
        "201":
          description: Pending payment submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSubmission"
        "400":
          description: Invalid payment submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payment-submissions/{id}:
    get:
      operationId: getPaymentSubmission
      tags: [payments]
      summary: Get payment confirmation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment submission status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSubmission"
        "404":
          description: Payment submission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/payment-submissions:
    get:
      operationId: listAdminPaymentSubmissions
      tags: [admin]
      summary: List payment submissions for admin review
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, paid, all]
            default: pending
      responses:
        "200":
          description: Payment submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentSubmission"
  /admin/payment-submissions/{id}/confirm:
    post:
      operationId: confirmAdminPaymentSubmission
      tags: [admin]
      summary: Confirm a payment submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Confirmed payment submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSubmission"
        "404":
          description: Payment submission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/drivers:
    get:
      operationId: listAdminDrivers
      tags: [drivers]
      summary: List driver accounts
      responses:
        "200":
          description: Driver accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createAdminDriver
      tags: [drivers]
      summary: Create a driver account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Created driver account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "400":
          description: Invalid driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/drivers/{id}:
    patch:
      operationId: updateAdminDriver
      tags: [drivers]
      summary: Update a driver account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdateInput"
      responses:
        "200":
          description: Updated driver account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "400":
          description: Invalid driver update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAdminDriver
      tags: [drivers]
      summary: Delete a driver account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted driver account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PaymentSubmission:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: string
        paymentMethod:
          type: string
        total:
          type: number
        transactionLast4:
          type: string
        status:
          type: string
          enum: [pending, paid]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - paymentMethod
        - total
        - transactionLast4
        - status
        - createdAt
        - updatedAt
    PaymentSubmissionInput:
      type: object
      properties:
        orderId:
          type: string
          minLength: 1
        paymentMethod:
          type: string
          minLength: 1
        total:
          type: number
          minimum: 0
        transactionLast4:
          type: string
          pattern: "^\\d{4}$"
      required:
        - orderId
        - paymentMethod
        - total
        - transactionLast4
    Driver:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicle:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - vehicle
        - status
        - createdAt
        - updatedAt
    DriverInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        vehicle:
          type: string
          minLength: 1
      required:
        - name
        - phone
        - vehicle
    DriverUpdateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        vehicle:
          type: string
          minLength: 1
        status:
          type: string
          enum: [active, inactive]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error