# lib/api-zod/src/generated/types/paymentSubmission.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentSubmissionStatus } from './paymentSubmissionStatus';

export interface PaymentSubmission {
  id: number;
  orderId: string;
  paymentMethod: string;
  total: number;
  transactionLast4: string;
  status: PaymentSubmissionStatus;
  createdAt: Date;
  updatedAt: Date;
}