# lib/api-zod/src/generated/types/paymentSubmissionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentSubmissionInput {
  /** @minLength 1 */
  orderId: string;
  /** @minLength 1 */
  paymentMethod: string;
  /** @minimum 0 */
  total: number;
  /** @pattern ^\d{4}$ */
  transactionLast4: string;
}